feat/page title in navbar
